refactor(http/prom): hoist `MkStreamLabel` out of `record_response`
`linkerd_http_prom::record_response` defines a pair of traits that are
used to derive labels for prometheus metrics. this is used to install
telemetry that inspects the response body, which is helpful for
recording response latency, and inspecting gRPC status codes.

this is a helpful model. to help reuse this trait, we hoist it out of
`record_response` and into a dedicated submodule.
