+++ linkerd/http/prom/src/lib.rs
@@ -4,3 +4,4 @@
 pub mod body_data;
 pub mod count_reqs;
 pub mod record_response;
+pub mod stream_label;
